feat(IAM Policy Management): add support for v2/policies API from IAM (#220)
* feat(IAM Policy Management): add support for v2/policies API from IAM
* new generated file based on API definition change, description added

+++ iampolicymanagementv1/iam_policy_management_v1_integration_test.go
@@ -25,6 +25,7 @@ import (
 	"math/rand"
 	"os"
 	"strconv"
+	"strings"
 	"time"
 
 	. "github.com/onsi/ginkgo"
@@ -47,7 +48,9 @@ var _ = Describe("IAM Policy Management - Integration Tests", func() {
 		testAccountID     string
 		etagHeader        string = "ETag"
 		testPolicyETag    string = ""
+		testV2PolicyETag    string = ""
 		testPolicyId      string = ""
+		testV2PolicyId    string = ""
 		testUserId        string = "IBMid-GoSDK" + strconv.Itoa(rand.Intn(100000))
 		testViewerRoleCrn string = "crn:v1:bluemix:public:iam::::role:Viewer"
 		testEditorRoleCrn string = "crn:v1:bluemix:public:iam::::role:Editor"
@@ -290,6 +293,223 @@ var _ = Describe("IAM Policy Management - Integration Tests", func() {
 		})
 	})
 
+	Describe("Create a v2 access policy", func() {
+
+		It("Successfully created a v2 access policy", func() {
+			shouldSkipTest()
+
+			// Construct an instance of the ResourceAttribute model
+			accountIdResourceAttribute := new(iampolicymanagementv1.V2PolicyAttribute)
+			accountIdResourceAttribute.Key = core.StringPtr("accountId")
+			accountIdResourceAttribute.Value = core.StringPtr(testAccountID)
+			accountIdResourceAttribute.Operator = core.StringPtr("stringEquals")
+
+			serviceNameResourceAttribute := new(iampolicymanagementv1.V2PolicyAttribute)
+			serviceNameResourceAttribute.Key = core.StringPtr("serviceType")
+			serviceNameResourceAttribute.Value = core.StringPtr("service")
+			serviceNameResourceAttribute.Operator = core.StringPtr("stringEquals")
+
+			// Construct an instance of the SubjectAttribute model
+			subjectAttribute := new(iampolicymanagementv1.V2PolicyAttribute)
+			subjectAttribute.Key = core.StringPtr("iam_id")
+			subjectAttribute.Operator = core.StringPtr("stringEquals")
+			subjectAttribute.Value = core.StringPtr(testUserId)
+
+			// Construct an instance of the V2PolicyBaseResource model
+			policyResource := new(iampolicymanagementv1.V2PolicyBaseResource)
+			policyResource.Attributes = []iampolicymanagementv1.V2PolicyAttribute{*accountIdResourceAttribute, *serviceNameResourceAttribute}
+
+			// Construct an instance of the PolicyRole model
+			policyRole := new(iampolicymanagementv1.PolicyRole)
+			policyRole.RoleID = core.StringPtr(testViewerRoleCrn)
+
+			// Construct an instance of the PolicySubject model
+			policySubject := new(iampolicymanagementv1.V2PolicyBaseSubject)
+			policySubject.Attributes = []iampolicymanagementv1.V2PolicyAttribute{*subjectAttribute}
+
+			// Contruct and instance of PolicyControl model
+			policyControl := new(iampolicymanagementv1.V2PolicyBaseControl)
+			controlGrant := new(iampolicymanagementv1.V2PolicyBaseControlGrant)
+			controlGrant.Roles = []iampolicymanagementv1.PolicyRole{*policyRole}
+			policyControl.Grant = controlGrant
+
+			// Construct an instance of Policy Rule Attribute
+			weeklyConditionAttribute := new(iampolicymanagementv1.V2PolicyAttribute)
+			weeklyConditionAttribute.Key = core.StringPtr("{{environment.attributes.day_of_week}}")
+			weeklyConditionAttribute.Operator = core.StringPtr("dayOfWeekAnyOf")
+			weeklyConditionAttribute.Value = []int{1, 2, 3, 4, 5}
+
+			startConditionAttribute := new(iampolicymanagementv1.V2PolicyAttribute)
+			startConditionAttribute.Key = core.StringPtr("{{environment.attributes.current_time}}")
+			startConditionAttribute.Operator = core.StringPtr("timeGreaterThanOrEquals")
+			startConditionAttribute.Value = core.StringPtr("09:00:00+00:00")
+
+			endConditionAttribute := new(iampolicymanagementv1.V2PolicyAttribute)
+			endConditionAttribute.Key = core.StringPtr("{{environment.attributes.current_time}}")
+			endConditionAttribute.Operator = core.StringPtr("timeLessThanOrEquals")
+			endConditionAttribute.Value = core.StringPtr("17:00:00+00:00")
+
+			policyRule := new(iampolicymanagementv1.V2PolicyBaseRule)
+			policyRule.Operator = core.StringPtr("and")
+			policyRule.Conditions = []iampolicymanagementv1.V2PolicyAttribute{*weeklyConditionAttribute, *startConditionAttribute, *endConditionAttribute}
+
+			// Construct an instance of the V2CreatePolicyOptions model
+			options := new(iampolicymanagementv1.V2CreatePolicyOptions)
+			options.Type = core.StringPtr("access")
+			options.Subject = policySubject
+			options.Control = policyControl
+			options.Resource = policyResource
+			options.Pattern = core.StringPtr("time-based-restrictions:weekly")
+			options.Rule = policyRule
+			options.AcceptLanguage = core.StringPtr("en")
+
+			policy, detailedResponse, err := service.V2CreatePolicy(options)
+			Expect(err).To(BeNil())
+			Expect(detailedResponse.StatusCode).To(Equal(201))
+			Expect(policy).ToNot(BeNil())
+			fmt.Fprintf(GinkgoWriter, "CreatePolicy() result:\n%s\n", common.ToJSON(policy))
+			Expect(policy.Type).To(Equal(options.Type))
+			Expect(policy.Subject.Attributes[0].Value).To(Equal(testUserId))
+			Expect(policy.Control.Grant.Roles[0].RoleID).To(Equal(options.Control.Grant.Roles[0].RoleID))
+			Expect(policy.Resource.Attributes[0].Value).To(Equal(testAccountID))
+
+			testV2PolicyId = *policy.ID
+		})
+	})
+
+	Describe("Get a v2 access policy", func() {
+
+		It("Successfully retrieved a v2 access policy", func() {
+			shouldSkipTest()
+			Expect(testPolicyId).To(Not(BeNil()))
+
+			options := service.NewV2GetPolicyOptions(testV2PolicyId)
+			policy, detailedResponse, err := service.V2GetPolicy(options)
+			Expect(err).To(BeNil())
+			Expect(detailedResponse.StatusCode).To(Equal(200))
+			Expect(policy).ToNot(BeNil())
+			fmt.Fprintf(GinkgoWriter, "GetPolicy() result:\n%s\n", common.ToJSON(policy))
+			Expect(*policy.ID).To(Equal(testV2PolicyId))
+
+			testV2PolicyETag = detailedResponse.GetHeaders().Get(etagHeader)
+		})
+	})
+
+	Describe("Update a v2 access policy", func() {
+
+		It("Successfully updated a v2 access policy", func() {
+			shouldSkipTest()
+			Expect(testV2PolicyId).To(Not(BeNil()))
+			Expect(testV2PolicyETag).To(Not(BeNil()))
+
+			// Construct an instance of the ResourceAttribute model
+			accountIdResourceAttribute := new(iampolicymanagementv1.V2PolicyAttribute)
+			accountIdResourceAttribute.Key = core.StringPtr("accountId")
+			accountIdResourceAttribute.Value = core.StringPtr(testAccountID)
+			accountIdResourceAttribute.Operator = core.StringPtr("stringEquals")
+
+			serviceNameResourceAttribute := new(iampolicymanagementv1.V2PolicyAttribute)
+			serviceNameResourceAttribute.Key = core.StringPtr("serviceType")
+			serviceNameResourceAttribute.Value = core.StringPtr("service")
+			serviceNameResourceAttribute.Operator = core.StringPtr("stringEquals")
+
+			// Construct an instance of the SubjectAttribute model
+			subjectAttribute := new(iampolicymanagementv1.V2PolicyAttribute)
+			subjectAttribute.Key = core.StringPtr("iam_id")
+			subjectAttribute.Operator = core.StringPtr("stringEquals")
+			subjectAttribute.Value = core.StringPtr(testUserId)
+
+			// Construct an instance of the V2PolicyBaseResource model
+			policyResource := new(iampolicymanagementv1.V2PolicyBaseResource)
+			policyResource.Attributes = []iampolicymanagementv1.V2PolicyAttribute{*accountIdResourceAttribute, *serviceNameResourceAttribute}
+
+			// Construct an instance of the PolicyRole model
+			policyRole := new(iampolicymanagementv1.PolicyRole)
+			policyRole.RoleID = core.StringPtr(testViewerRoleCrn)
+
+			// Construct an instance of the PolicySubject model
+			policySubject := new(iampolicymanagementv1.V2PolicyBaseSubject)
+			policySubject.Attributes = []iampolicymanagementv1.V2PolicyAttribute{*subjectAttribute}
+
+			// Contruct and instance of PolicyControl model
+			policyControl := new(iampolicymanagementv1.V2PolicyBaseControl)
+			controlGrant := new(iampolicymanagementv1.V2PolicyBaseControlGrant)
+			controlGrant.Roles = []iampolicymanagementv1.PolicyRole{*policyRole}
+			policyControl.Grant = controlGrant
+
+			// Construct an instance of Policy Rule Attribute
+			weeklyConditionAttribute := new(iampolicymanagementv1.V2PolicyAttribute)
+			weeklyConditionAttribute.Key = core.StringPtr("{{environment.attributes.day_of_week}}")
+			weeklyConditionAttribute.Operator = core.StringPtr("dayOfWeekAnyOf")
+			weeklyConditionAttribute.Value = []int{1, 2, 3, 4, 5}
+
+			startConditionAttribute := new(iampolicymanagementv1.V2PolicyAttribute)
+			startConditionAttribute.Key = core.StringPtr("{{environment.attributes.current_time}}")
+			startConditionAttribute.Operator = core.StringPtr("timeGreaterThanOrEquals")
+			startConditionAttribute.Value = core.StringPtr("09:00:00+00:00")
+
+			endConditionAttribute := new(iampolicymanagementv1.V2PolicyAttribute)
+			endConditionAttribute.Key = core.StringPtr("{{environment.attributes.current_time}}")
+			endConditionAttribute.Operator = core.StringPtr("timeLessThanOrEquals")
+			endConditionAttribute.Value = core.StringPtr("17:00:00+00:00")
+
+			policyRule := new(iampolicymanagementv1.V2PolicyBaseRule)
+			policyRule.Operator = core.StringPtr("and")
+			policyRule.Conditions = []iampolicymanagementv1.V2PolicyAttribute{*weeklyConditionAttribute, *startConditionAttribute, *endConditionAttribute}
+
+			// Construct an instance of the UpdatePolicyOptions model
+			options := new(iampolicymanagementv1.V2UpdatePolicyOptions)
+			options.PolicyID = core.StringPtr(testV2PolicyId)
+			options.IfMatch = core.StringPtr(testV2PolicyETag)
+			options.Type = core.StringPtr("access")
+			options.Subject = policySubject
+			options.Control = policyControl
+			options.Resource = policyResource
+			options.Pattern = core.StringPtr("time-based-restrictions:weekly")
+			options.Rule = policyRule
+
+			policy, detailedResponse, err := service.V2UpdatePolicy(options)
+			Expect(err).To(BeNil())
+			Expect(detailedResponse.StatusCode).To(Equal(200))
+			Expect(policy).ToNot(BeNil())
+			fmt.Fprintf(GinkgoWriter, "UpdatePolicy() result:\n%s\n", common.ToJSON(policy))
+			Expect(*policy.ID).To(Equal(testV2PolicyId))
+			Expect(policy.Type).To(Equal(options.Type))
+			Expect(policy.Subject.Attributes[0].Value).To(Equal(testUserId))
+			Expect(policy.Control.Grant.Roles[0].RoleID).To(Equal(options.Control.Grant.Roles[0].RoleID))
+			Expect(policy.Resource.Attributes[0].Value).To(Equal(testAccountID))
+
+			newV2PolicyEtag := detailedResponse.GetHeaders().Get(etagHeader)
+			Expect(newV2PolicyEtag).ToNot(Equal(testV2PolicyETag))
+
+		})
+	})
+
+	Describe("List v2 access policies", func() {
+
+		It("Successfully listed the account's v2 access policies", func() {
+			shouldSkipTest()
+			Expect(testV2PolicyId).To(Not(BeNil()))
+
+			options := service.NewV2ListPoliciesOptions(testAccountID)
+			options.SetIamID(testUserId)
+			result, detailedResponse, err := service.V2ListPolicies(options)
+			Expect(err).To(BeNil())
+			Expect(detailedResponse.StatusCode).To(Equal(200))
+			Expect(result).ToNot(BeNil())
+			fmt.Fprintf(GinkgoWriter, "ListPolicies() result:\n%s\n", common.ToJSON(result))
+
+			// confirm the test policy is present
+			testPolicyPresent := false
+			for _, policy := range result.Policies {
+				if *policy.ID == testV2PolicyId {
+					testPolicyPresent = true
+				}
+			}
+			Expect(testPolicyPresent).To(BeTrue())
+		})
+	})
+
 	Describe("Create custom role", func() {
 		It("Successfully created custom role", func() {
 			shouldSkipTest()
@@ -399,12 +619,20 @@ var _ = Describe("IAM Policy Management - Integration Tests", func() {
 				continue
 			}
 			fiveMinutesAgo := time.Now().Add(-(time.Duration(5) * time.Minute))
-
-			if *policy.ID == testPolicyId || createdAt.Before(fiveMinutesAgo) {
-				options := service.NewDeletePolicyOptions(*policy.ID)
-				detailedResponse, err := service.DeletePolicy(options)
-				Expect(err).To(BeNil())
-				Expect(detailedResponse.StatusCode).To(Equal(204))
+			if strings.Contains(*policy.Href, "v2/policies") {
+				if *policy.ID == testV2PolicyId || createdAt.Before(fiveMinutesAgo) {
+					options := service.NewV2DeletePolicyOptions(*policy.ID)
+					detailedResponse, err := service.V2DeletePolicy(options)
+					Expect(err).To(BeNil())
+					Expect(detailedResponse.StatusCode).To(Equal(204))
+				}
+			} else {
+				if *policy.ID == testPolicyId || createdAt.Before(fiveMinutesAgo) {
+					options := service.NewDeletePolicyOptions(*policy.ID)
+					detailedResponse, err := service.DeletePolicy(options)
+					Expect(err).To(BeNil())
+					Expect(detailedResponse.StatusCode).To(Equal(204))
+				}
 			}
 		}
